+++ www/go/base/db/ActiveRecord.php
@@ -1499,6 +1499,8 @@ abstract class ActiveRecord extends \GO\Base\Model{
 				$where .= ') ';
 			} else
 			{
+				go()->setOptimizerSearchDepth();
+
 				$joins .= "\nINNER JOIN core_search search ON search.entityId = t.id and search.entityTypeId = " . static::entityType()->getId();
 
 				$i = 0;