+++ www/go/core/App.php
@@ -803,6 +803,30 @@ use const GO_CONFIG_FILE;
 
 			$blob->output(true);
 		}
+
+		private $optimizerSearchDepthSet = false;
+
+		/**
+		 * Used for global search queries to optimize execution plan
+		 * Queries could hang in 'Statistics' state when inputing a large string resulting in many joins.
+		 *
+		 * Also see: https://mariadb.com/resources/blog/setting-optimizer-search-depth-in-mysql/
+		 */
+		public function setOptimizerSearchDepth() {
+
+			if($this->optimizerSearchDepthSet) {
+				return;
+			}
+
+			try {
+				go()->getDbConnection()->exec("SET SESSION optimizer_search_depth=4;");
+
+			} catch(Exception $e) {
+				ErrorHandler::log("Could not set 'optimizer_search_depth' on mysql. Please configure manually. " . $e->getMessage());
+			}
+
+			$this->optimizerSearchDepthSet = true;
+		}
 	}
 
 }